% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumexp_functions.R
\name{integrate}
\alias{integrate}
\title{Vectorized integrate function in \code{lower}, \code{upper}}
\usage{
integrate(
  f,
  lower,
  upper,
  ...,
  subdivisions = 100L,
  rel.tol = .Machine$double.eps^0.25,
  abs.tol = rel.tol,
  stop.on.error = TRUE,
  keep.xy = FALSE,
  aux = NULL
)
}
\description{
Vectorized integrate function in \code{lower}, \code{upper}
}
