% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumexp_functions.R
\name{lebesgue_integral}
\alias{lebesgue_integral}
\title{Lebesgue Integration}
\usage{
lebesgue_integral(fnx, base_measure = list(Lebesgue = c(0, Inf)), from, to)
}
\arguments{
\item{fnx}{The function to be measured}

\item{base_measure}{A named list giving the type of measure to be used, with elements specifying the domain
of integration. Measures include
\itemize{
\item Lebesgue: The standard measure; defaults to Riemann integration using the base R \code{integrate} approximation
\item Atomic: A point mass; the value of the function at the point mass is added
\item Simple: A measure over a specified number of breaks, expecting a right-continuous step-function over those intervals
in the breaks.
\item \<User\>: Any function defined on the domain of integration in \code{from} and \code{to}. For instance, dwei = c(0, 60) will
integrate \code{fnx} against the Weibull probability density function using \code{integrate}.
Note: The domain for atomic masses must be length 1 or length
of \code{from} or \code{to}
}}

\item{from, to}{The set over which the function should be measured (wrt the base measure)}
}
\description{
Adds atomic masses and other measures to \code{integrate}
}
