% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumexp_functions.R
\name{matrix.constr}
\alias{matrix.constr}
\title{Design Matrix by Size}
\usage{
matrix.constr(group_sizes)
}
\description{
Construct a design matrix by giving a vector of group_sizes (avoids \code{cut} for simple-function construction)
}
