% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumexp_functions.R
\name{sm_epsilon}
\alias{sm_epsilon}
\title{a small number
a small number for use in calculating approximate function values on open intervals}
\usage{
sm_epsilon(power = -10)
}
\description{
a small number
a small number for use in calculating approximate function values on open intervals
}
