% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumexp_functions.R
\name{stress_function}
\alias{stress_function}
\title{Stress Function}
\usage{
stress_function(
  step_domain = numeric(),
  shock_domain = step_domain,
  stress = numeric(length(step_domain + 1)),
  xi = numeric()
)
}
\arguments{
\item{step_domain}{Numeric value giving the point at which the stress should
increase.}

\item{shock_domain}{Numeric value giving the point at which the instantaneous
shock \code{xi} should occur; defaults to \code{step_domain}}

\item{stress}{Numeric vector giving the value of the stresses, must have 1 more
unit than the step_domain}

\item{xi}{The value of the shock stress, should be same length 1 or same length
as \code{shock_domain}}

\item{mode}{The type of stress function, constant stress gives models equivalent to
their marginal distribution; step-stress gives models with variable stresses
over time.}

\item{shock}{Logical indicating if the model should include a \code{xi} parameter
governinng the probability of instant failure at the shock_domain}
}
\value{
An object of class "stress_function" which includes the function (x$fnx)
and any specified atoms of shock stress (x$atom = shock_domain).
}
\description{
A function giving the idealized stress a unit operates under over time.
}
